v1.4.10 ~替换大量默认规则为自写规则（acl4ssr优化版） ~增加新手引导中对开启公网访问的指引 ~增加defaultDNS参数（无需额外配置） ~修复部分情况下配置文件报错bug ~尝试适配meta核心的tun模块（未测试） ~增加自动下载meta核心必需的geosite.dat数据库功能

+++ scripts/start.sh
@@ -641,7 +641,7 @@ bfstart(){
 			setconfig Geo_v $Geo_v
 		fi
 	fi
-	if [ ! -f $bindir/geosite.dat ];then
+	if [ "$clashcore" = "clash.meta" -a ! -f $bindir/geosite.dat ];then
 		if [ -f $clashdir/geosite.dat ];then
 			mv $clashdir/geosite.dat $bindir/geosite.dat
 		else
